pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "pgpt"
version = "0.0.4"
description = "ProgramEngineerGPT is an interactive command line tool that leverages the power of AI to assist developers with code comprehension, exploration, and generation. It serves as a virtual assistant that can analyze codebases, answer queries about code, and even help in setting up new coding projects."
authors = [
    { name = "Brette Geary", email = "hackedbg@outlook.com" },
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    'aioboto3==11.2.0',
    'aiobotocore==2.5.0',
    'aiohttp==3.8.4',
    'aioitertools==0.11.0',
    'aiosignal==1.3.1',
    'async-timeout==4.0.2',
    'attrs==23.1.0',
    'black==23.3.0',
    'boto3==1.26.76',
    'botocore==1.29.76',
    'certifi==2023.5.7',
    'charset-normalizer==3.1.0',
    'click==8.1.3',
    'colored==2.1.1',
    'dataclasses-json==0.5.8',
    'deeplake==3.6.6',
    'dill==0.3.6',
    'entrypoints==0.4',
    'exceptiongroup==1.1.1',
    'frozenlist==1.3.3',
    'greenlet==2.0.2',
    'humbug==0.3.1',
    'idna==3.4',
    'iniconfig==2.0.0',
    'isort==5.12.0',
    'jmespath==1.0.1',
    'langchain==0.0.212',
    'langchainplus-sdk==0.0.17',
    'marshmallow==3.19.0',
    'marshmallow-enum==1.5.1',
    'multidict==6.0.4',
    'multiprocess==0.70.14',
    'mypy-extensions==1.0.0',
    'nest-asyncio==1.5.6',
    'numcodecs==0.11.0',
    'numexpr==2.8.4',
    'numpy==1.25.0',
    'openai==0.27.8',
    'openapi-schema-pydantic==1.2.4',
    'packaging==23.1',
    'pathos==0.3.0',
    'pathspec==0.11.1',
    'Pillow==9.5.0',
    'platformdirs==3.8.0',
    'pluggy==1.2.0',
    'pox==0.3.2',
    'ppft==1.7.6.6',
    'prompt-toolkit==3.0.38',
    'pydantic==1.10.9',
    'PyJWT==2.7.0',
    'pytest==7.4.0',
    'python-dateutil==2.8.2',
    'python-dotenv==1.0.0',
    'PyYAML==6.0',
    'questionary==1.10.0',
    'regex==2023.6.3',
    'requests==2.31.0',
    'ruff==0.0.275',
    's3transfer==0.6.1',
    'six==1.16.0',
    'SQLAlchemy==2.0.17',
    'tenacity==8.2.2',
    'tiktoken==0.4.0',
    'tomli==2.0.1',
    'tqdm==4.65.0',
    'typing-inspect==0.9.0',
    'typing_extensions==4.6.3',
    'urllib3==1.26.16',
    'wcwidth==0.2.6',
    'wrapt==1.15.0',
    'yarl==1.9.2',
]

[tool.setuptools]
packages = ["programengineergpt"]

[project.scripts]
pgpt = 'programengineergpt.main:run'

[project.urls]
"Homepage" = "https://github.com/hackedbyagirl/programengineergpt"
"Bug Tracker" = "https://github.com/hackedbyagirl/programengineergpt/issues"

[tool.ruff]
select = ["F", "E", "W", "I001"]
line-length = 90
show-fixes = false
target-version = "py311"
task-tags = ["TODO", "FIXME"]
exclude = [
 ".bzr",
 ".direnv",
 ".eggs",
 ".git",
 ".ruff_cache",
 ".svn",
 ".tox",
 ".venv",
 "__pypackages__",
 "_build",
 "buck-out",
 "build",
 "dist",
 "node_modules",
 "venv",
]

[tool.ruff.isort]
known-first-party = []
known-third-party = []
section-order = [
 "future",
 "standard-library",
 "third-party",
 "first-party",
 "local-folder",
]
combine-as-imports = true
split-on-trailing-comma = false
lines-between-types = 1

[tool.black]
line-length = 90
target-version = ["py311"]
include = '\\.pyi?$'
exclude = '''
(
 /(
  .direnv
  | .eggs
  | .git
  | .tox
  | .venv
  | _build
  | build
  | dist
  | venv
 )/
)
'''
